move common modals folder
