+++ src/common/modals/base.vue
@@ -30,7 +30,7 @@ export default class ModalBase extends Vue {
 </script>
 
 <style scoped lang="scss">
-@import '../../../colors';
+@import '../../colors';
 
 .modal {
   position: absolute;
